+++ src/lib/components/table/holders/index.tsx
@@ -1,0 +1,4 @@
+export * from "./HoldersTable";
+export * from "./HoldersTableHeader";
+export * from "./HoldersTableMobileCard";
+export * from "./HoldersTableRow";